[filtermod] Merge Cisco ACI/APIC into Cisco IOS
The Splunk add-on for IOS also supports APIC and ACL the previous source type configuration was incorrect
