+++ deploy/k8s-microk8s/sc4s-infra.yaml
@@ -42,8 +42,6 @@ data:
     checkpoint_splunk_web,index,netproxy
     checkpoint_splunk,index,netops
     checkpoint_splunk,index,netops
-    cisco_apic_acl,index,netfw
-    cisco_apic_events,index,netops
     cisco_acs,index,netauth
     cisco_asa,index,netfw
     cisco_ftd,index,netfw
