+++ package/etc/context_templates/splunk_metadata.csv.example
@@ -15,8 +15,6 @@ checkpoint_splunk,index,netops
 checkpoint_splunk,index,netops
 checkpoint_splunk,index,netops
 cisco_acs,index,netauth
-cisco_apic_acl,index,netfw
-cisco_apic_events,index,netops
 cisco_ace,index,netops
 cisco_asa,index,netfw
 Cisco_C100V Email Security Virtual Appliance_ESA_CONSOLIDATED_LOG_EVENT,index,email
